// packages/frontend-2/src/generated/typechain/CampaignFactory.ts
export declare namespace Campaign {
  export type SharesDataStruct = {
    totalShares: BigNumberish;
    sharesMerkleRoot: BytesLike;
  };

  export type SharesDataStructOutput = [BigNumber, string] & {
    totalShares: BigNumber;
    sharesMerkleRoot: string;
  };
}

export interface CampaignFactoryInterface extends utils.Interface {
  contractName: "CampaignFactory";
  functions: {
    "createCampaign((uint256,bytes32),bytes32,address,address,bool,uint256,bytes32)": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "createCampaign",
    values: [
      Campaign.SharesDataStruct,
      BytesLike,
      string,
      string,
      boolean,
      BigNumberish,
      BytesLike
    ]
  ): string;

  decodeFunctionResult(
    functionFragment: "createCampaign",
    data: BytesLike
  ): Result;

  events: {
    "CampaignCreated(address,address,bytes32,uint256,bytes32,address,address,bool,uint256,bytes32)": EventFragment;
  };

  getEvent(nameOrSignatureOrTopic: "CampaignCreated"): EventFragment;
}

export type CampaignCreatedEvent = TypedEvent<
  [
    string,
    string,
    string,
    BigNumber,
    string,
    string,
    string,
    boolean,
    BigNumber,
    string
  ],
  {
    creator: string;
    newCampaign: string;
    _sharesRoot: string;
    _sharesTotal: BigNumber;
    _uri: string;
    _guardian: string;
    _oracle: string;
    _sharesPublished: boolean;
    _claimPeriodStart: BigNumber;
    salt: string;
  }
>;

export type CampaignCreatedEventFilter = TypedEventFilter<CampaignCreatedEvent>;

export interface CampaignFactory extends BaseContract {
  contractName: "CampaignFactory";
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: CampaignFactoryInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    createCampaign(
      _shares: Campaign.SharesDataStruct,
      _uri: BytesLike,
      _guardian: string,
      _oracle: string,
      _sharesPublished: boolean,
      _claimPeriodStart: BigNumberish,
      salt: BytesLike,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<ContractTransaction>;
  };

  createCampaign(
    _shares: Campaign.SharesDataStruct,
    _uri: BytesLike,
    _guardian: string,
    _oracle: string,
    _sharesPublished: boolean,
    _claimPeriodStart: BigNumberish,
    salt: BytesLike,
    overrides?: Overrides & { from?: string | Promise<string> }
  ): Promise<ContractTransaction>;

  callStatic: {
    createCampaign(
      _shares: Campaign.SharesDataStruct,
      _uri: BytesLike,
      _guardian: string,
      _oracle: string,
      _sharesPublished: boolean,
      _claimPeriodStart: BigNumberish,
      salt: BytesLike,
      overrides?: CallOverrides
    ): Promise<void>;
  };

  filters: {
    "CampaignCreated(address,address,bytes32,uint256,bytes32,address,address,bool,uint256,bytes32)"(
      creator?: null,
      newCampaign?: null,
      _sharesRoot?: null,
      _sharesTotal?: null,
      _uri?: null,
      _guardian?: null,
      _oracle?: null,
      _sharesPublished?: null,
      _claimPeriodStart?: null,
      salt?: null
    ): CampaignCreatedEventFilter;
    CampaignCreated(
      creator?: null,
      newCampaign?: null,
      _sharesRoot?: null,
      _sharesTotal?: null,
      _uri?: null,
      _guardian?: null,
      _oracle?: null,
      _sharesPublished?: null,
      _claimPeriodStart?: null,
      salt?: null
    ): CampaignCreatedEventFilter;
  };

  estimateGas: {
    createCampaign(
      _shares: Campaign.SharesDataStruct,
      _uri: BytesLike,
      _guardian: string,
      _oracle: string,
      _sharesPublished: boolean,
      _claimPeriodStart: BigNumberish,
      salt: BytesLike,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    createCampaign(
      _shares: Campaign.SharesDataStruct,
      _uri: BytesLike,
      _guardian: string,
      _oracle: string,
      _sharesPublished: boolean,
      _claimPeriodStart: BigNumberish,
      salt: BytesLike,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<PopulatedTransaction>;
  };
}